Issue of negative alphas
Addresses an issue that some alphas were violating constraints.
This is an issue of the QP solver osqp. According to https://github.com/oxfordcontrol/osqp/issues/151 it comes from the parameter adaptive_rho that, when set to TRUE (the default) prevent convergence.
It is thus only set to FALSE by default in my code to prevent such an issue.

In addition, the functions related to alpha are separated in another file for better organization of functions.